+++ main.js
@@ -11,6 +11,7 @@ import './module/scene-music.js';
 import './module/end-credits.js';
 import './module/dramatic-health-display.js';
 import './module/playback-sync.js';
+import './module/journal-transfer.js';
 import { migrate } from './module/migrate.js';
 import('./module/encounter.js');